fix(vigilantia): Casus-Modal kompilierbar — .to(int) für Cluster-Badge
rx.foreach liefert event-Items als Var[Any] (ObjectItemOperation),
direkte int-Vergleiche (``> 1``) und Arithmetik (``- 1``) sprengen den
Reflex-Compiler. Cast auf .to(int) löst den TypeError beim Compile.

Hatte den selben Fehler-Modus früher schon einmal beim
motion_min_area_ratio-Layout-Patch im Vorschau-Popup.

+++ aifred/ui/casus.py
@@ -207,11 +207,12 @@ def _event_row(event: rx.Var) -> rx.Component:
         _thumb(event),
         _event_type_badge(event),
         # Cluster-Member-Badge — nur sichtbar im Cluster-Mode wenn der
-        # Cluster mehr als 1 Mitglied hat. Tooltip zeigt Cluster-ID.
+        # Cluster mehr als 1 Mitglied hat. Reflex Var[Any] kann nicht
+        # direkt mit int verglichen werden → .to(int) Cast.
         rx.cond(
-            event["cluster_member_count"] > 1,
+            event["cluster_member_count"].to(int) > 1,
             rx.badge(
-                "+" + (event["cluster_member_count"] - 1).to(str),
+                "+" + (event["cluster_member_count"].to(int) - 1).to(str),
                 color_scheme="purple",
                 variant="soft",
                 size="1",
